removing trailing white space
removing fixture

fixing changes from code review
