+++ cypress/integration-qe/memberUserCreateCluster.spec.js
@@ -1,13 +1,13 @@
+import { CLUSTER_NAME } from './shared';
 import {
-  createClusterHosted,
-  loginOCM,
-  verifyClusterCreation,
-  CLUSTER_NAME,
   OCM_USER,
   OCM_USER_PASSWORD,
-} from './shared';
+  loginOCM,
+  createClusterHosted,
+  verifyClusterCreationHosted,
+} from './ocmShared';
 
-describe('Member User Can Created Cluster in hosted env', () => {
+describe('member user can create cluster in hosted env', () => {
   it('log in to OCM', () => {
     // Set CYPRESS_OCM_USER and CYPRESS_OCM_USER_PASSWORD env variables
     loginOCM(OCM_USER, OCM_USER_PASSWORD);
@@ -15,6 +15,6 @@ describe('Member User Can Created Cluster in hosted env', () => {
 
   it('Can navigate to assisted-installer and create cluster', () => {
     createClusterHosted(CLUSTER_NAME);
-    verifyClusterCreation(CLUSTER_NAME);
+    verifyClusterCreationHosted(CLUSTER_NAME);
   });
 });
